requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
pdf2image